continue try ns
