+++ lib/config.py
@@ -83,7 +83,7 @@ wrong_relation_num = 999999999999  # 错误的关系，设置9999可以是全部
 total_questions = 99999999
 stop_loss_zeor_count = 2000  # 2000次则停下来
 rnn_size = 100
-mode = "cc"
+mode = "cc" #cc 中文训练  cc_test 中文测试
 check = 100000
 # 属性模式 # 1 num 限制数量 2 special 指定 3 no 非训练模式 4 maybe 模糊属性的单独处理
 use_property = 'special'  # 使用属性的模式做训练和测试
@@ -97,7 +97,7 @@ batch_size = 5  # 1个batch的大小 # 临时改成1 个看loss
 
 # ==================================需要配置
 # mark = '测试CP的效果；寻找最佳的CP策略；10P;NEG负例的个数是全部 ;14-cp的竞争策略[20181013-1]' # 备注
-mark = '测试CP的效果；100p;竞争CP；螺旋式上升' # 备注
+mark = '测试CP的效果；' # 备注
 t_relation_num = 100  # 4385  # 重要！这个指示了训练的关系个数 4358
 ner_top_cand = 0  # 训练取2，测试取3（写死） ; 0 只测属性识别,2 测实体或者实体+属性
 
@@ -108,13 +108,12 @@ loss_part = '...'  # entity_relation ；entity_relation_transE;entity;relation
 #  IR-GAN
 ns_model = 'competing_q'  # competing_q_ert competing_q  only_default only_default_er  random entity temp_test_all
 g_epoches = 0
-d_epoches = 2
+d_epoches = 1
 s_epoches = 0
 c_epoches = 0
 a_epoches = 0
 ner_epoches = 0
 ns_epoches = 1
-
 
 # ===============================实验的时候才调整
 batch_size_gan = 200  # 100 或者 1000 ,80%的竞争属性是在600
@@ -123,22 +122,29 @@ sampled_temperature = 20
 gan_learn_rate = 0.05
 d_need_cal_attention = True
 g_need_cal_attention = True
-competing_s_neg_p_num = 10   # 竞争属性中，P_POS的负例的多少
+competing_s_neg_p_num = 10   # 竞争属性中，P_POS的负例的多少 用于 ert
 competing_p_pos_neg_size = 9999  # 竞争属性中，P_POS的负例的多少
 convert_rs_to_words = False   # 关系集合转换为对应的字符集合
 only_p_neg_in_cp = False               # 只加入P_NEG进入 CP 暂停，这样会导致POS无法加入
-hand_add_some_neg = True
+hand_add_some_neg = False
 does_cp_contains_default = False # competing_q 模式下 是否包含默认的 属性。去掉才显得出
 # only_default 默认模式|fixed_amount 固定 最多100个 | additional 默认+额外
 # synonym_train_mode 优先加入neg的同义词
 # competing_ps 竞争属性
 # pool_mode = 'additional' # competing_ps | None | additional | only_default
 
+# 一些调整的参数
+ns_ps_try_only_pos = False
+ns_ps_len_max_limit = 22
+ns_q_ploicy_all = 'all_p' # D的策略，全问题还是一个 all_p , 1_q,1_p
 # 模型恢复
 restore_model = True
 restore_path_base = r'F:\PycharmProjects\dl2\deeplearning\QA_GAN\runs'
-restore_path = restore_path_base+\
-                 r'\2018_10_28_15_16_00_100p_cp_2\checkpoints\step=2_epoches=d_index=0\model.ckpt-1'
+restore_path = restore_path_base+ \
+               r'\2018_11_05_20_13_54_100p_retest\checkpoints\step=1_epoches=d_index=0\model.ckpt-1'
+                 # r'\2018_10_26_10_58_29_100p_default\checkpoints\step=3_epoches=d_index=0\model.ckpt-1'
+#                  r'\2018_11_01_11_45_50_allp_default_1\checkpoints\step=1_epoches=d_index=0\model.ckpt-1'
+# r'\2018_11_05_16_45_17_100p_cp_4\checkpoints\step=21_epoches=d_index=0\model.ckpt-1'
                 #r'\2018_10_05_11_06_59_allp_ns_cp_best\checkpoints\step=6_epoches=d_index=0\model.ckpt-1'
 restore_test = False
 
@@ -148,7 +154,7 @@ restore_test = False
 competing_ps_path = '../data/nlpcc2016/14-cp/competing_ps_tj.v2.txt'
 # competing_ps_path = '../data/nlpcc2016/13-competing/competing_ps_tj.v2.txt'# 13-competing 版本的
 # competing_ps_path = '../data/nlpcc2016/13-competing/competing_p_in_kb.v2.txt'
-competing_batch_size = 10 # 控制size
+competing_batch_size = 10 # 控制size 无用
 
 expend_es = '../data/nlpcc2016/4-ner/result/q.rdf.score.top_3_all_0.v4.10.txt'
 
@@ -266,7 +272,10 @@ cc_p = {
     'convert_rs_to_words':convert_rs_to_words,
     'only_p_neg_in_cp':only_p_neg_in_cp,
     'hand_add_some_neg':hand_add_some_neg,
-    'does_cp_contains_default':does_cp_contains_default
+    'does_cp_contains_default':does_cp_contains_default,
+    'ns_ps_len_max_limit':ns_ps_len_max_limit,
+    'ns_ps_try_only_pos':ns_ps_try_only_pos,
+    'ns_q_ploicy_all':ns_q_ploicy_all
 
 }
 
@@ -331,6 +340,7 @@ ms = ["train", "test"
     , "debug_epoches"
     , "bad"
     , "loss"
+      ,'cost'
       # expection
       ]
